Added PWM freq timer for PDM
if any of the motor PWMs are in a certain threshold range, initialize startup mode by adding the desired time to the current time. Startup mode will run until the current time is greater than the time set by the initializer. only writes the frequency to the teeny's PWM timers if they change.